+++ docs/development/README.md
@@ -90,6 +90,13 @@ pnpm build
 
 随后即可调试修改数据定义包、核心包或者其它代码。
 
+### 关于运行时
+
+由于项目使用 TypeScript 编写，由于包含历史代码、生成代码等原因，项目多数无法直接简单转译为 JavaScript（如 `enum`、构造函数自动参数、不带扩展名的 `import`、legacy 装饰器等），故项目使用预加载 `tsx` 的 `node` 来运行这些代码（从而不用在开发时进行额外转译构建步骤）。在 `@gi-tcg/config` 中定义了 `gnx` 脚本（auGmented Node eXecution）来包装 `node` 以支持 TypeScript 模块解析和转译，之后该入口也将为 DSL 转译做准备。
+
+- 在 `scripts` 中，使用 `gnx` 来替代 `node`，如: `gnx scripts/foo.ts`；
+- 在命令行中，使用 `pnpm gnx`。
+
 ### 例：启动 `@gi-tcg/standalone` 项目的开发服务器
 
 ```
