+++ packages/assets-manager/package.json
@@ -27,7 +27,7 @@
     }
   },
   "scripts": {
-    "build:data": "nod --import=./scripts/fetch_preload.ts ./scripts/fetch.ts",
+    "build:data": "gnx --import=./scripts/fetch_preload.ts ./scripts/fetch.ts",
     "build:ts": "tsdown",
     "build": "pnpm build:data && pnpm build:ts"
   }
