+++ packages/cbinding/package.json
@@ -21,7 +21,7 @@
     "rollup": "^4.27.4"
   },
   "scripts": {
-    "build": "nod scripts/prepare.ts",
+    "build": "gnx scripts/prepare.ts",
     "cmake:configure": "cmake -B build -GNinja",
     "cmake:build": "cmake --build build --config Release",
     "cmake:install": "cmake --install build --config Release --prefix install"
