+++ packages/data-code-analyzer/package.json
@@ -5,7 +5,7 @@
   "main": "src/index.ts",
   "types": "src/index.ts",
   "scripts": {
-    "build": "nod scripts/index.ts"
+    "build": "gnx scripts/index.ts"
   },
   "devDependencies": {
     "@gi-tcg/data": "workspace:*"
